Figure out how sforzando VST saves its settings. Fix paths.
The settings are encoded with zlib (after a header). First decompressing
the relevant part, then search and replacing the paths and finally
compressing the decompressed and modified xml produces the correct file.
Some changes needed for header (encode compressed and decompressed
length).

Seems that the decompressed part is in ARIA format which essentially is
xml.

The file paths are also in unix format e.g. directory paths are
separated with the "/" char.

Now fixing the paths also for sforzando.